+++ controls/aem_ready.rb
@@ -33,3 +33,21 @@ control 'publish-dispatcher-instances-ready' do
     it { should have_all_publish_dispatcher_instances_ready }
   end
 end
+
+control 'orchestrator-instances-ready' do
+  impact 1.0
+  title 'Check if Orchestrator instances are ready'
+  desc 'Check if Orchestrator instances are ready'
+  describe ready do
+    it { should have_all_orchestrator_instances_ready }
+  end
+end
+
+control 'chaosmonkey-instances-ready' do
+  impact 1.0
+  title 'Check if Chaosmonkey instances are ready'
+  desc 'Check if Chaosmonkey instances are ready'
+  describe ready do
+    it { should have_all_chaosmonkey_instances_ready }
+  end
+end