+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.11.7
-*
+* Disable recovery of publish instance #16
 
 ### 0.11.6
 * Add new checks for un/scheduling live-snapshots shinesolutions/aem-aws-stack-builder#212