+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.11.7
 * Add feature to terminate author-primary instance
+* Add readiness check for AEM OpenCloud components Chaos Monkey & Orchestrator
+* Add recovery testing for AEM OpenCloud components Chaos Monkey & Orchestrator
 
 ### 0.11.6
 * Add new checks for un/scheduling live-snapshots shinesolutions/aem-aws-stack-builder#212